Add Nav.js test
